im: Move VectorSubscriber into a separate module
